src/requirements.txt
earthengine-api
requests
python-dotenv
pandas
psycopg2-binary
